Only admin can enable/disable users

+++ app/Http/Controllers/AdminController.php
@@ -12,26 +12,37 @@ class AdminController extends Controller
 
     public function listAllUsers()
     {
-        $users = User::all();
+        $users = User::whereNotIn('id', [2])->get();
         return view('admin')->with (compact('users'));
     }
 
-    public function enable($id): RedirectResponse
+    public function enable($id,Request $request): RedirectResponse
     {
-        $user = User::find($id);
-        $user->is_enabled = true;
-        $user->save();
+        if($request->user()->is_admin == 1){
+            $user = User::find($id);
+            $user->is_enabled = true;
+            $user->save();
 
-        return redirect()->back();
+            return redirect()->back();
+        }else{
+           return redirect('home')->with('error', 'You don\'t have admin access.');
+        }
+
     }
 
-    public function disable($id): RedirectResponse
+    public function disable($id, Request $request): RedirectResponse
     {
-        $user = User::find($id);
-        $user->is_enabled = false;
-        $user->save();
-        $user->sessions()->delete();
+        if($request->user()->is_admin == 1){
+            $user = User::find($id);
+            $user->is_enabled = false;
+            $user->save();
+            $user->sessions()->delete();
 
-        return redirect()->back();
+            return redirect()->back();
+        }else{
+            return redirect()->route('home')->with('error', 'You don\'t have admin access.');
+        }
+
+
     }
 }
